requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.52.2
anyio==3.7.1
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
black==23.12.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.8
coloredlogs==15.0.1
Deprecated==1.2.18
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.0
fastapi==0.104.1
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.5.1
google-auth==2.40.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.72.1
h11==0.16.0
hf-xet==1.1.2
hiredis==3.2.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
jiter==0.10.0
kubernetes==32.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.0.2
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.82.1
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
overrides==7.7.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
posthog==4.2.0
propcache==0.3.1
protobuf==5.29.5
pulsar-client==3.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.19.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==7.4.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
PyYAML==6.0.2
redis==5.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
structlog==23.2.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.22.0
